requirement.txt
slackclient
cassandra-driver
